+++ packages/terra-clinical-detail-view/src/terra-dev-site/doc/example/DetailViewDividedSmallerTitles.jsx
@@ -62,6 +62,7 @@ const DetailViewDivided = () => (
   <div className={cx('detail-view-divided')}>
     <DetailView
       title="Multiple Ingredients"
+      level={4}
       secondaryTitles={['M. V. I. Adult 10 mL = 10 mL', 'Sodium bicarbonate 8.4% 50 mL', 'Dextrose 5% in Water 1000 mL']}
       subtitles={['30 mg, Oral, Start Date/Time: 06/28/17 20:00:00 CDT. Stop Date/Time: 06/28/17 20:00:00 CDT']}
       accessory={(